Allow compilation on osx

+++ build.sh
@@ -1,28 +1,34 @@
-#!/bin/bash
+#!/usr/bin/env bash -e
+
+set -o pipefail
 
 # Simple packaging of zookeepercli
 #
 # Requires fpm: https://github.com/jordansissel/fpm
 #
 
+platform=$(uname -s)
 release_version="1.0.10"
 release_dir=/tmp/zookeepercli
-rm -rf $release_dir/*
+rm -rf ${release_dir:?}/*
 mkdir -p $release_dir
 
-cd  $(dirname $0)
+pushd "$(dirname "$0")"
 for f in $(find . -name "*.go"); do go fmt $f; done
 
-go build -o $release_dir/zookeepercli
+go get github.com/outbrain/golib/log || exit 1
+go get github.com/samuel/go-zookeeper/zk || exit 1
+go build -o $release_dir/zookeepercli || exit 1
 
-if [[ $? -ne 0 ]] ; then
-	exit 1
+if [ "$platform" = "Linux" ]; then
+  pushd "$release_dir"
+  # rpm packaging
+  fpm -v "${release_version}" -f -s dir -t rpm -n zookeepercli -C $release_dir --prefix=/usr/bin .
+  fpm -v "${release_version}" -f -s dir -t deb -n zookeepercli -C $release_dir --prefix=/usr/bin .
+  popd
 fi
-
-cd $release_dir
-# rpm packaging
-fpm -v "${release_version}" -f -s dir -t rpm -n zookeepercli -C $release_dir --prefix=/usr/bin .
-fpm -v "${release_version}" -f -s dir -t deb -n zookeepercli -C $release_dir --prefix=/usr/bin .
 
 echo "---"
 echo "Done. Find releases in $release_dir"
+
+popd
